+++ src/neunorm/__init__.py
@@ -24,7 +24,9 @@ try:
 except ImportError:
     __version__ = "unknown"
 
+from neunorm.data_models.roi import ROI
+
 __author__ = "Jean Bilheux, Chen Zhang"
 __email__ = "bilheuxjm@ornl.gov, zhangc@ornl.gov"
 
-__all__ = ["__version__"]
+__all__ = ["ROI", "__version__"]